:truck: move metadata helper interfaces
