+++ lib/helpers/index.ts
@@ -1,13 +1,5 @@
 export * from './default-command-publisher';
 export * from './default-event-serializer';
 export * from './default-query-publisher';
-export * from './get-aggregate-metadata';
-export * from './get-command-handler-metadata';
-export * from './get-command-metadata';
-export * from './get-event-listener-metadata';
-export * from './get-event-metadata';
-export * from './get-event-serializer-metadata';
-export * from './get-query-handler-metadata';
-export * from './get-query-metadata';
-export * from './get-snapshot-metadata';
+export * from './metadata';
 export * from './observable-bus';
