+++ lib/helpers/metadata/get-command-handler-metadata.ts
@@ -1,5 +1,5 @@
-import { COMMAND_HANDLER_METADATA } from '../decorators';
-import { CommandHandlerMetadata } from '../interfaces';
+import { COMMAND_HANDLER_METADATA } from '../../decorators';
+import { CommandHandlerMetadata } from '../../interfaces';
 
 export const getCommandHandlerMetadata = (commandHandler: Function): CommandHandlerMetadata => {
 	return Reflect.getMetadata(COMMAND_HANDLER_METADATA, commandHandler) ?? {};
