+++ lib/helpers/metadata/get-event-listener-metadata.ts
@@ -1,5 +1,5 @@
-import { EVENT_LISTENER_METADATA } from '../decorators';
-import { EventListenerMetadata } from '../interfaces';
+import { EVENT_LISTENER_METADATA } from '../../decorators';
+import { EventListenerMetadata } from '../../interfaces';
 
 export const getEventListenerMetadata = (eventListener: Function): EventListenerMetadata => {
 	return Reflect.getMetadata(EVENT_LISTENER_METADATA, eventListener) ?? {};
