+++ lib/helpers/metadata/get-event-metadata.ts
@@ -1,6 +1,6 @@
 import { Type } from '@nestjs/common';
-import { EVENT_METADATA } from '../decorators';
-import { EventMetadata, IEvent } from '../interfaces';
+import { EVENT_METADATA } from '../../decorators';
+import { EventMetadata, IEvent } from '../../interfaces';
 
 export const getEventMetadata = (event: Type<IEvent>): EventMetadata => {
 	return Reflect.getMetadata(EVENT_METADATA, event) ?? {};
