+++ lib/helpers/metadata/get-query-handler-metadata.ts
@@ -1,5 +1,5 @@
-import { QUERY_HANDLER_METADATA } from '../decorators';
-import { QueryHandlerMetadata } from '../interfaces';
+import { QUERY_HANDLER_METADATA } from '../../decorators';
+import { QueryHandlerMetadata } from '../../interfaces';
 
 export const getQueryHandlerMetadata = (queryHandler: Function): QueryHandlerMetadata => {
 	return Reflect.getMetadata(QUERY_HANDLER_METADATA, queryHandler) ?? {};
